feat(play): use game start room on load
Remove hardcoded start room id and use the one set for the game.
