+++ app/game/[gameId]/play/page.tsx
@@ -1,95 +1,37 @@
-"use client";
+import { getGameById } from "@/features/games/services";
+import { getRoomById } from "@/features/rooms/services";
+import { PlayView } from "./play-view";
+import { BadgeAlert } from "lucide-react";
 
-import { useParams } from "next/navigation";
-import { useEffect, useMemo, useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { getRoom } from "@/features/rooms/api";
-import { Direction, Exit, Room } from "@/features/rooms/types";
-
-const INTIIAL_ROOM_ID = 4;
-
-export default function Play() {
-  const params = useParams<{ gameId: string }>();
-  const gameId = useMemo(() => Number.parseInt(params.gameId), [params.gameId]);
-  const [room, setRoom] = useState<Room | null>(null);
-  const [command, setCommand] = useState("");
-
-  useEffect(() => {
-    const initialRoomId = INTIIAL_ROOM_ID;
-
-    async function fetchRoom() {
-      try {
-        const nextRoom = await getRoom(gameId, initialRoomId);
-        setRoom(nextRoom);
-      } catch (error) {
-        console.error(error);
-      }
-    }
-
-    fetchRoom();
-  }, [gameId]);
-
-  async function handleMove(direction: Direction) {
-    if (room === null) return;
-    const exit: Exit = `${direction}Exit`;
-    if (room[exit] === null) {
-      alert(`There is no room to the ${direction}!`);
-    } else {
-      try {
-        const nextRoom = await getRoom(gameId, room[exit]);
-        setRoom(nextRoom);
-      } catch (error) {
-        alert((error as Error).message);
-      }
-    }
+export default async function PlayPage({
+  params,
+}: {
+  params: Promise<{ gameId: string }>;
+}) {
+  const { gameId } = await params;
+  const game = await getGameById(Number.parseInt(gameId));
+  if (game === undefined) {
+    return <div>No game found</div>;
   }
 
-  function handleCommand(command: string) {
-    if (room === null) return;
-
-    const normalizedCommand = command.toLowerCase();
-    switch (normalizedCommand) {
-      case "move north":
-        handleMove("north");
-        break;
-      case "move south":
-        handleMove("south");
-        break;
-      case "move east":
-        handleMove("east");
-        break;
-      case "move west":
-        handleMove("west");
-        break;
-    }
-  }
-
-  function handleRun() {
-    handleCommand(command);
-    setCommand("");
-  }
-
-  return (
-    <div className="grid h-full w-full place-items-center">
-      <div className="relative flex w-xl flex-col items-start gap-4 overflow-hidden rounded-md p-4 pl-6 shadow before:absolute before:top-0 before:left-0 before:h-full before:w-2 before:bg-emerald-600">
-        <div className="text-2xl font-semibold tracking-tight">Play Game</div>
-        <div>You are in: {room?.name ?? "Loading..."}</div>
-        <div>
-          <Label className="pb-2" htmlFor="command">
-            Enter command
-          </Label>
-          <span className="flex flex-row gap-2">
-            <Input
-              id="command"
-              value={command}
-              onChange={(e) => setCommand(e.target.value)}
-            />
-            <Button onClick={handleRun}>Run</Button>
-          </span>
+  if (game.startRoomId === null) {
+    return (
+      <div className="grid h-full w-full place-items-center">
+        <div className="relative flex w-xl flex-col items-start gap-4 overflow-hidden rounded-md p-4 pl-6 shadow before:absolute before:top-0 before:left-0 before:h-full before:w-2 before:bg-emerald-600">
+          <div className="text-4xl font-bold tracking-tight">Play Game</div>
+          <div className="relative flex flex-row items-center gap-2 overflow-hidden rounded-sm bg-amber-200 p-2 pl-4 before:absolute before:left-0 before:h-full before:w-2 before:bg-amber-400">
+            <BadgeAlert className="size-7 text-amber-500" />
+            <span className="font-medium">No start room has been set!</span>
+          </div>
         </div>
       </div>
-    </div>
-  );
+    );
+  }
+
+  const startRoom = await getRoomById(game.id, game.startRoomId);
+  if (startRoom === undefined) {
+    return <div>Invalid start room ID</div>;
+  }
+
+  return <PlayView startRoom={startRoom} />;
 }
